+++ app/Http/Controllers/Api/TaskController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\Task;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -16,9 +17,12 @@ class TaskController extends Controller
      * Display a listing of the resource.
      *
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('viewAny', Task::class);
+
         return response()->json(Task::all(), Response::HTTP_OK);
     }
 
@@ -28,10 +32,13 @@ class TaskController extends Controller
      * @param Request $request
      * @return JsonResponse
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
-        $data = $this->validate($request, Task::rules());
+        $this->authorize('create', Task::class);
+
+        $data = $this->validate($request, Task::createRules());
 
         $task = Task::create($data);
 
@@ -43,9 +50,12 @@ class TaskController extends Controller
      *
      * @param  Task  $task
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function show(Task $task)
     {
+        $this->authorize('view', $task);
+
         return response()->json($task, Response::HTTP_OK);
     }
 
@@ -56,10 +66,13 @@ class TaskController extends Controller
      * @param Task $task
      * @return JsonResponse
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function update(Request $request, Task $task)
     {
-        $data = $this->validate($request, Task::rules());
+        $this->authorize('update', $task);
+
+        $data = $this->validate($request, Task::updateRules());
 
         $task->update($data);
 
@@ -72,9 +85,12 @@ class TaskController extends Controller
      * @param Task $task
      * @return JsonResponse
      * @throws Exception
+     * @throws AuthorizationException
      */
     public function destroy(Task $task)
     {
+        $this->authorize('delete', $task);
+
         $users = $task->users();
         $user_ids = $users->allRelatedIds()->all();
         $users->detach($user_ids);
